refactor: split router.rs into handlers/{auth,user,session}
