+++ backend-v2/src/handlers/mod.rs
@@ -1,0 +1,3 @@
+pub mod auth;
+pub mod session;
+pub mod user;
