+++ backend-v2/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod auth;
 pub mod error;
+pub mod handlers;
 pub mod model;
 pub mod repository;
 pub mod router;
